新增充值详情备注字段

+++ app/src/main/java/com/zhifeng/kuangchi/module/CarryDetailDto.java
@@ -61,7 +61,7 @@ public class CarryDetailDto {
          */
 
         private int id;
-        private int money;
+        private double money;
         private String address;
         private int user_id;
         private int coin_type;
@@ -69,7 +69,7 @@ public class CarryDetailDto {
         private String fronzen_money;
         private long add_time;
         private String tax_rate;
-        private int input_money;
+        private double input_money;
         private String tax;
 
         public int getId() {
@@ -80,11 +80,11 @@ public class CarryDetailDto {
             this.id = id;
         }
 
-        public int getMoney() {
+        public double getMoney() {
             return money;
         }
 
-        public void setMoney(int money) {
+        public void setMoney(double money) {
             this.money = money;
         }
 
@@ -144,11 +144,11 @@ public class CarryDetailDto {
             this.tax_rate = tax_rate == null ? "" : tax_rate;
         }
 
-        public int getInput_money() {
+        public double getInput_money() {
             return input_money;
         }
 
-        public void setInput_money(int input_money) {
+        public void setInput_money(double input_money) {
             this.input_money = input_money;
         }
